+++ lib/jobmon/railtie.rb
@@ -2,7 +2,6 @@ module Jobmon
   class Railtie < ::Rails::Railtie
     rake_tasks do
       require 'jobmon/tasks/preload'
-      load 'jobmon/tasks/test.rake'
       load 'jobmon/tasks/delayed_job_queue_monitor.rake'
       load 'jobmon/tasks/sidekiq_queue_monitor.rake'
     end