+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.4.0
 
 - `jobmon`コマンド追加
+- `task_with_monitor` DSL を非推奨に
+- `jobmon:test_job` タスク削除
 
 ## 0.3.0